Way to spawn dragon
Changelog from first commit:
Made it remove the "root_of_continuity" block from the player inv, works in stack sizes

Added a way to reset the Holster so that it can be used again. Done by using a "netherrite knife" on the "Alive" Holster, which drops "haemopal"

Reformatted things into their own functions to make it easy to read.

TODO:
Make the "Dead" Holster not drop anything,
make the "Alive" Holster match the properties of the "Dead" Holster (sound type and material stuff)

possible improvements:
make creative work so that items are not taken and tools are not damaged